requirements.txt
Django>=3.2,<5.0
python-telegram-bot>=20.0,<21.0
django-ninja>=0.17.0
python-dotenv>=1.0.0
requests>=2.0.0
djangorestframework>=3.12.0
sniffio>=1.3.0
anyio>=3.7.0
httpx>=0.24.0
aiohttp>=3.8.0
